api: buffer: Change buffer size from size_t to uint32_t
Use uint32_t for buffer size to be align with packet API lenghts
and offsets.

+++ platform/linux-generic/include/api/odp_buffer.h
@@ -44,7 +44,7 @@ void *odp_buffer_addr(odp_buffer_t buf);
  *
  * @return Buffer maximum data size
  */
-size_t odp_buffer_size(odp_buffer_t buf);
+uint32_t odp_buffer_size(odp_buffer_t buf);
 
 /**
  * Buffer type
